+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.0.4] - [Github repo]
+### Fix
+- Serial is reachable fix
 ## [0.0.3] - 2020-02-24
 ### Added
 - Monokai style to jupyter console